Minor enhancement: forced a compile time conversion of GRAVITY value to an integer, rather than at run time.
git-svn-id: https://gentlenav.googlecode.com/svn/trunk@368 cf4e4d4a-e611-11dd-8fb1-9960f2a117f8

+++ MatrixPilot/servoOut.c
@@ -130,7 +130,7 @@ void __attribute__((__interrupt__,__no_auto_psv__)) _PWMInterrupt(void)
 		xrate.offset = xrate.value ;
 		yaccel.offset = yaccel.value ;
 		yrate.offset = yrate.value ;
-		zaccel.offset = zaccel.value - 2*GRAVITY ; // GRAVITY is measured in A-D/2 units
+		zaccel.offset = zaccel.value - ((int)(2*GRAVITY)) ; // GRAVITY is measured in A-D/2 units
 		zrate.offset = zrate.value ;
 #ifdef VREF
 		vref.offset = vref.value ;
